203-Add 10 point gain in rule adoption
Adds in the ability to gain points when a rule is accepted. While this is attempted in rule 202, along with the original nomic rules, it appears to not be entered correctly allowing for no point gain. 
Instead of amending rule 202, it is more cohesive to modify rule 203 which determines the criteria for rule acceptance. This makes gaining points automatic after the rule acceptance instead of a determined 'turn' step.