Create 390.md
Rationale: given that the number of players is very small (currently 12 players,) a scenario in which a very small number of players arrive at a meeting is plausible. Example: 2 players (aside from the Professor) arrive at a meeting on a harsh weather day. Voting shouldn't take place in such a case. Even though the players may choose the majority (7) to be the minimum number, this number must still be added to the set of rules.